+++ src/Config/Config.php
@@ -29,7 +29,6 @@ class Config extends ActiveRecordConfig {
     const KEY_ENABLE_USERS_VIEW = "enable_users_view";
     const KEY_REPORTING_ALWAYS_SHOW_CHILD_TYPES = "reporting_always_show_child_types";
     const KEY_SHOW_MATRIX_ACTIONS = "show_matrix_actions";
-    const KEY_SHOW_ONLY_APPEARABLE_ORG_UNITS_IN_FILTER = "show_only_appearable_org_units_in_filter";
     const KEY_SYNC_POSITION_PERMISSIONS_WITH_CHILDREN = "sync_position_permissions_with_children";
 	/**
 	 * @var array
@@ -46,7 +45,6 @@ class Config extends ActiveRecordConfig {
         self::KEY_ENABLE_USERS_VIEW => [ self::TYPE_BOOLEAN, false ],
         self::KEY_REPORTING_ALWAYS_SHOW_CHILD_TYPES => [ self::TYPE_JSON, [] ],
         self::KEY_SHOW_MATRIX_ACTIONS => [ self::TYPE_BOOLEAN, false ],
-        self::KEY_SHOW_ONLY_APPEARABLE_ORG_UNITS_IN_FILTER => [self::TYPE_BOOLEAN, false ],
         self::KEY_SYNC_POSITION_PERMISSIONS_WITH_CHILDREN => [self::TYPE_BOOLEAN, false ]
 	];
 }